chore: Move extension to separate file
