+++ src/Uno.UI.Runtime.Skia.Tizen/Tizen/ApplicationModel/TizenPackageIdExtension.cs
@@ -1,6 +1,7 @@
 ﻿#nullable enable
 
 using Tizen.Applications;
+using Uno.ApplicationModel;
 using Windows.ApplicationModel;
 using Package = Tizen.Applications.Package;
