+++ src/Uno.UWP/ApplicationModel/IPackageIdExtension.skia.cs
@@ -1,0 +1,15 @@
+﻿using Windows.ApplicationModel;
+
+namespace Uno.ApplicationModel
+{
+	internal interface IPackageIdExtension
+	{
+		string FamilyName { get; }
+
+		string FullName { get; }
+
+		string Name { get; }
+
+		PackageVersion Version { get; }
+	}
+}
